% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{circGenomePlot}
\alias{circGenomePlot}
\title{Generation of a circular genomic plot}
\usage{
circGenomePlot(x, gff = gff, proteinID = proteinID, reorigff = FALSE)
}
\arguments{
\item{x}{DNAStringSet of nucleotide sequences in fasta formate.}

\item{gff}{Processed gff file, has to be output of either functions processProkkagff() or processNCBIgff(), since a column with alternative end is generated that is used in this function. This file can have been reoriented afterwards with function reorientgff().}

\item{proteinID}{Vector of ProteinIDs to be indicated in the plot.}

\item{reorientgff}{If TRUE uses columns with names "Ostart" and "Oend" to obtain the base pair location. Defaults to FALSE and uses columns with names "start" and "end".}
}
\value{
List containing the circular plot and data.frame of regulator location.
}
\description{
Generates a genomic plot indicating locations of regulators and showing the GC skew, based on gff file.
only one replicon.
}
\examples{
vect <- c("AAV93333.1", "AAV93335.1")
plot <- circGenomePlot(fasta, gff3, vect, reorigff = TRUE)
}
