% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gff_convert}
\alias{gff_convert}
\title{Processing of Prokka Gff output}
\usage{
gff_convert(x, reorient = FALSE)
}
\arguments{
\item{x}{directory and file name of the Gff object to be processed. File could be read with function list.files.object.}

\item{reorient}{should function reorientGff() be used after this function = TRUE}
}
\description{
A function that processes Gff files that originate from Prokka output.
}
\examples{
gff_convert()
}
\keyword{Prokka}
\keyword{gff}
