% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{processNCBIgff}
\alias{processNCBIgff}
\title{Processing of a gff file type downloaded from NCBI}
\usage{
processNCBIgff(x)
}
\arguments{
\item{x}{The unprocessed gff file downloaded from NCBI.}
}
\value{
The processed gff file as data.frame with an additional column containing the alternative end, used for function(circGenomePlot)
}
\description{
Process the gff file provided by NCBI into a data.frame.
}
\examples{
gff_unprocessed <- path_to_gff
gff <- processNCBIgff(gff_unprocessed);
}
