% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reorientFasta}
\alias{reorientFasta}
\title{Reorientation of fasta files}
\usage{
reorientFasta(x, replicon = 1, gff = NA, location = position)
}
\arguments{
\item{x}{list of DNAStringSets}

\item{replicon}{the replicon to be reoriented. Defaults to the largest replicon. Numbers increase with degreasing length (largest replicon = 1, second largest = 2, ...).}

\item{gff}{list of tables containing gff information. Required if location is given by Protein ID.}

\item{location}{vector of the starting positions in base pairs at which the Fasta file is to begin. Alternatively, a vector of protein IDs can be used if the corresponding gff objects are supplied.}
}
\description{
A function that batch reorientes a list of fasta files to start at the location given as vector of bp position or Protein IDs.
}
\examples{
reorientFasta()
}
\keyword{genomics}
