% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reorientfna}
\alias{reorientfna}
\title{Reorientate a genome fasta file}
\usage{
reorientfna(x, replicon = NA, bplocation = NA, proteinID = NA, gff = NA)
}
\arguments{
\item{x}{DNAStringSet of nucleotide sequences in fasta formate.}

\item{replicon}{Replicon to be reoriented. This option defaults to the largest replicon.}

\item{bplocation}{Location in base pairs that should be used as new start position.}

\item{proteinID}{ProteinID indicating the protein based on which the file should be reoriented. Must be accompanied by a processed gff file and either be located on the largest replicon or also be accompanied by an indication of the replicon to be used.}

\item{gff}{A processed gff file (e.g. using functions processProkkagff or processNCBIgff). Must be supplied when option ProteinID is selected.}
}
\value{

}
\description{
Adjust the start position of a fna file downloaded from NCBI.
}
\examples{
fasta <- reorientfna(x = dna_list, proteinID = "AAV93333.1", gff = gff3)
}
