% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reorientGff}
\alias{reorientGff}
\title{Reorientation of gff files}
\usage{
reorientGff(x, location = position, replicon = NA, prokka = FALSE)
}
\arguments{
\item{x}{list of tables containing gff information.}

\item{location}{starting positions in base pairs or of protein IDs.}

\item{replicon}{replicon identifiers based on which the Gff files should be subsetted.}

\item{prokka}{did the gff file originate as prokka output.}
}
\description{
A function that batch reorientes a list of gff files to start at the location given as vector of bp position or Protein IDs.
}
\examples{
reorientGff()
}
\keyword{genomics}
