% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{circGenomePlot}
\alias{circGenomePlot}
\title{Circular genomic plots}
\usage{
circGenomePlot(
  x,
  gff = NA,
  proteinID = proteinID,
  radius1 = 10,
  radius2 = 12,
  radius3 = 13,
  radius4 = 14,
  radius5 = 15
)
}
\arguments{
\item{x}{list of DNAStringSets.}

\item{gff}{list of tables containing gff information. Files need to be output of function reorientGff() or colnames of start and end that should be used need to be labeled "NewStart" and "NewEnd".}

\item{proteinID}{vector of protein IDs whose gene position is to be shown in the plot.}

\item{radius}{radius adjustment of each circle.}
}
\description{
A function that plots the GC Skew, genes on plus and minus strand and location of selected genes. The output consists of a list containing the plot and a list of the regulators with their locations.
}
\examples{
circGenomePlot()
}
\keyword{genomics}
\keyword{plot}
